requirements.txt
Django==1.10.2
flake8
djangorestframework